Fix preview mouse position on Windows
1. Fix preview mouse position on Windows
On Windows, if the screen resize is not 1, the device resolution might
be different from the mouse resolution.

The "touch_down" and "touch_up" event use the device resolution,
but the "on_mouse_pos" event use the mouse resolution. So the mouse
position is not correct. We need to convert the mouse position to
the device resolution.

2. Improve performance of find preview.